# exports/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "description": "返回所有用户基础信息，仅管理员可访问。",
        "operationId": "UsersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "用户数组",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "用户列表（admin 可见）",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "创建普通用户或管理员账号，密码会进行哈希存储。",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建后的用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建用户（admin）",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "按用户 ID 查询单个用户信息。",
        "operationId": "UsersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户 ID",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查看用户",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me": {
      "delete": {
        "description": "当前用户自销；内置账号禁止删除。",
        "operationId": "UsersController_deleteSelf",
        "parameters": [],
        "responses": {
          "200": {
            "description": "被删除的用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "自销账号（内置账号不可销）",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/templates": {
      "get": {
        "description": "返回模板规则的最新版本，用于前端选择启用。",
        "operationId": "TemplatesController_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "模板列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateRuleSetVersionResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "列出所有模板最新版本",
        "tags": [
          "Templates"
        ]
      }
    },
    "/templates/{templateName}/versions": {
      "get": {
        "description": "用于模板版本回溯与回滚。",
        "operationId": "TemplatesController_listByName",
        "parameters": [
          {
            "name": "templateName",
            "required": true,
            "in": "path",
            "description": "模板名称",
            "schema": {
              "example": "ip_whitelist",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "版本列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateRuleSetVersionResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查看指定模板的历史版本",
        "tags": [
          "Templates"
        ]
      },
      "post": {
        "description": "提交模板规则并生成新版本。",
        "operationId": "TemplatesController_create",
        "parameters": [
          {
            "name": "templateName",
            "required": true,
            "in": "path",
            "description": "模板名称",
            "schema": {
              "example": "ip_whitelist",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "新版本信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateRuleSetVersionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建模板新版本",
        "tags": [
          "Templates"
        ]
      }
    },
    "/templates/{templateName}/versions/{versionNo}/rollback": {
      "post": {
        "description": "复制目标版本生成新版本，保留历史。",
        "operationId": "TemplatesController_rollback",
        "parameters": [
          {
            "name": "templateName",
            "required": true,
            "in": "path",
            "description": "模板名称",
            "schema": {
              "example": "ip_whitelist",
              "type": "string"
            }
          },
          {
            "name": "versionNo",
            "required": true,
            "in": "path",
            "description": "目标版本号",
            "schema": {
              "example": 2,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "回滚后的新版本",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateRuleSetVersionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "回滚模板到指定版本（生成新版本）",
        "tags": [
          "Templates"
        ]
      }
    },
...
    "/servers/{serverName}/publish": {
      "post": {
        "description": "组合策略 JSON 并写入规则文件，更新 nginx.conf 后执行 `nginx -t -q` 与 reload；失败会回滚并返回步骤日志。",
        "operationId": "RulesController_publish",
        "parameters": [
          {
            "name": "serverName",
            "required": true,
            "in": "path",
            "description": "站点名称（nginx server_name）",
            "schema": {
              "example": "example1.com",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishPolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发布结果（结构化 steps 与 publishLog）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishPolicyResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "下发并创建新策略版本（生成规则文件 → 更新 nginx.conf → 校验 → reload）",
        "tags": [
          "ServerPolicy"
        ]
      }
    },
    "/servers/{serverName}/runtime": {
      "put": {
        "description": "基于最新策略生成新版本并更新 nginx.conf，失败时回滚；返回结构化步骤输出。",
        "operationId": "RulesController_updateRuntime",
        "parameters": [
          {
            "name": "serverName",
            "required": true,
            "in": "path",
            "description": "站点名称（nginx server_name）",
            "schema": {
              "example": "example1.com",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuntimeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "运行态更新结果（steps/rollbackSteps）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeUpdateResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改运行态开关（wafEnabled/dynamicBlock/defaultAction），带锁与失败回滚",
        "tags": [
          "ServerPolicy"
        ]
      }
    },
...
    "/reports/waf": {
      "get": {
        "description": "返回防护报告摘要数据，用于前端图表渲染。",
        "operationId": "ReportsController_getReport",
        "parameters": [],
        "responses": {
          "200": {
            "description": "报告摘要",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WafReportSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取防护报告（JSON 摘要）",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/waf/export": {
      "get": {
        "description": "导出当前报告摘要的 PDF 文件。",
        "operationId": "ReportsController_exportPdf",
        "parameters": [],
        "responses": {
          "200": {
            "description": "PDF 文件",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "导出防护报告 PDF",
        "tags": [
          "Reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "WAF 控制面 API",
    "description": "v1 接口文档",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户 ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "admin"
          },
          "role": {
            "type": "string",
            "description": "角色",
            "enum": [
              "admin",
              "user"
            ]
          },
          "displayName": {
            "type": "string",
            "description": "显示名称",
            "example": "Admin"
          },
          "builtIn": {
            "type": "boolean",
            "description": "是否内置账号",
            "example": true
          }
        },
        "required": [
          "id",
          "username",
          "role",
          "builtIn"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "alice"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "123456"
          },
          "role": {
            "type": "string",
            "description": "角色",
            "enum": [
              "admin",
              "user"
            ]
          },
          "displayName": {
            "type": "string",
            "description": "显示名称",
            "example": "Alice"
          }
        },
        "required": [
          "username",
          "password",
          "role"
        ]
      },
...
      "TemplateRuleSetVersionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "版本 ID",
            "example": 12
          },
          "templateName": {
            "type": "string",
            "description": "模板名称",
            "example": "ip_whitelist"
          },
          "versionNo": {
            "type": "number",
            "description": "版本号",
            "example": 3
          },
          "rulesJson": {
            "type": "object",
            "description": "模板规则 JSON",
            "additionalProperties": true
          },
          "status": {
            "type": "string",
            "description": "状态",
            "enum": [
              "SUCCESS",
              "FAILED",
              "ROLLED_BACK"
            ]
          },
          "note": {
            "type": "string",
            "description": "备注"
          },
          "createdBy": {
            "type": "string",
            "description": "创建人",
            "example": "admin"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间（ISO）"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间（ISO）"
          }
        },
        "required": [
          "id",
          "templateName",
          "versionNo",
          "rulesJson",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTemplateDto": {
        "type": "object",
        "properties": {
          "templateName": {
            "type": "string",
            "description": "模板名称",
            "example": "ip_whitelist"
          },
          "rules": {
            "description": "规则列表（friendly JSON 结构）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleDto"
            }
          },
          "note": {
            "type": "string",
            "description": "备注",
            "example": "自定义 IP 白名单模板"
          },
          "actor": {
            "type": "string",
            "description": "操作者",
            "example": "alice"
          }
        },
        "required": [
          "templateName",
          "rules"
        ]
      },
      "RollbackTemplateDto": {
        "type": "object",
        "properties": {
          "actor": {
            "type": "string",
            "description": "操作者",
            "example": "alice"
          }
        }
      },
...
      "PublishPolicyDto": {
        "type": "object",
        "properties": {
          "enabledCoreRules": {
            "description": "启用的核心规则集名称列表（一级引用）",
            "example": [
              "core_sqli_rules"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enabledTemplates": {
            "description": "启用的模板名称列表（一级引用）",
            "example": [
              "ip_whitelist"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rules": {
            "description": "自定义规则数组（按 friendly-waf-json-spec 子集，必填字段：id/target/match/pattern/action）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleDto"
            }
          },
          "note": {
            "type": "string",
            "description": "版本备注",
            "example": "上线核心规则+自定义IP黑名单"
          },
          "dynamicBlockBaseScore": {
            "type": "number",
            "description": "动态封禁基础得分（policies.dynamicBlock.baseAccessScore）",
            "example": 1,
            "minimum": 0
          },
          "dryRun": {
            "type": "boolean",
            "description": "仅测试，不落盘/不reload",
            "default": false
          },
          "baseVersionId": {
            "type": "number",
            "description": "基线版本号（用于并发冲突校验）",
            "minimum": 1,
            "example": 12
          },
          "actor": {
            "type": "string",
            "description": "操作者",
            "example": "alice"
          }
        },
        "required": [
          "enabledCoreRules",
          "enabledTemplates"
        ]
      },
...
      "PublishPolicyResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "发布状态",
            "enum": [
              "SUCCESS",
              "FAILED"
            ]
          },
          "version": {
            "type": "number",
            "description": "新版本号（失败时可能为空）"
          },
          "dryRun": {
            "type": "boolean",
            "description": "是否为 dryRun",
            "example": false
          },
          "publishLog": {
            "type": "string",
            "description": "发布日志摘要（含步骤输出）"
          },
          "steps": {
            "description": "执行步骤",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineStepResultDto"
            }
          },
          "rollbackSteps": {
            "description": "回滚步骤（若有）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineStepResultDto"
            }
          },
          "error": {
            "type": "string",
            "description": "失败原因（仅失败时）"
          }
        },
        "required": [
          "status",
          "dryRun",
          "publishLog",
          "steps",
          "rollbackSteps"
        ]
      },
      "UpdateRuntimeDto": {
        "type": "object",
        "properties": {
          "wafEnabled": {
            "type": "boolean",
            "description": "是否启用 WAF（入口开关，true=启用，false=旁路）",
            "example": true
          },
          "dynamicBlockEnabled": {
            "type": "boolean",
            "description": "是否启用动态封禁（true=打开动态信誉/封禁）",
            "example": true
          },
          "defaultAction": {
            "type": "string",
            "description": "默认动作（DENY/BYPASS/ALLOW），空则保持现状",
            "example": "DENY"
          },
          "actor": {
            "type": "string",
            "description": "操作者",
            "example": "ops-user"
          },
          "note": {
            "type": "string",
            "description": "备注",
            "example": "紧急旁路 30 分钟"
          }
        }
      },
      "RuntimeUpdateResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "发布状态",
            "enum": [
              "SUCCESS",
              "FAILED"
            ]
          },
          "version": {
            "type": "number",
            "description": "新版本号（失败时可能为空）"
          },
          "publishLog": {
            "type": "string",
            "description": "发布日志摘要（含步骤输出）"
          },
          "steps": {
            "description": "执行步骤",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineStepResultDto"
            }
          },
          "rollbackSteps": {
            "description": "回滚步骤（若有）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineStepResultDto"
            }
          },
          "error": {
            "type": "string",
            "description": "失败原因（仅失败时）"
          }
        },
        "required": [
          "status",
          "publishLog",
          "steps",
          "rollbackSteps"
        ]
      },
...
      "AttackBreakdownDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "攻击类型",
            "example": "SQL_INJECTION"
          },
          "count": {
            "type": "number",
            "description": "数量"
          }
        },
        "required": [
          "type",
          "count"
        ]
      },
...
      "WafReportSummaryDto": {
        "type": "object",
        "properties": {
          "timeRange": {
            "type": "string",
            "description": "统计时间范围，如 5m/1h/24h",
            "example": "24h"
          },
          "generatedAt": {
            "type": "string",
            "description": "生成时间 ISO 字符串",
            "example": "2026-01-03T01:23:45Z"
          },
          "kpis": {
            "description": "核心指标",
            "allOf": [
              {
                "$ref": "#/components/schemas/WafReportKpiDto"
              }
            ]
          },
          "timeline": {
            "description": "请求/拦截时序（用于折线图）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeSeriesPointDto"
            }
          },
          "attackTypes": {
            "description": "攻击类型分布",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttackBreakdownDto"
            }
          },
          "geoWorld": {
            "description": "世界热点",
            "allOf": [
              {
                "$ref": "#/components/schemas/GeoSectionDto"
              }
            ]
          },
          "geoChina": {
            "description": "国内热点",
            "allOf": [
              {
                "$ref": "#/components/schemas/GeoSectionDto"
              }
            ]
          },
          "topUrls": {
            "description": "高频 URL 榜单",
            "example": [
              {
                "name": "/login",
                "count": 120
              },
              {
                "name": "/api/upload",
                "count": 35
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeoEntryDto"
            }
          },
          "topAttackIps": {
            "description": "高频攻击 IP 榜单",
            "example": [
              {
                "name": "1.1.1.1",
                "count": 30
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeoEntryDto"
            }
          },
          "topBlockedUrls": {
            "description": "高频拦截 URL 榜单",
            "example": [
              {
                "name": "/admin",
                "count": 12
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeoEntryDto"
            }
          },
          "topBlockedIps": {
            "description": "高频拦截 IP 榜单",
            "example": [
              {
                "name": "2.2.2.2",
                "count": 7
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeoEntryDto"
            }
          }
        },
        "required": [
          "timeRange",
          "generatedAt",
          "kpis",
          "timeline",
          "attackTypes",
          "geoWorld",
          "geoChina",
          "topUrls",
          "topAttackIps",
          "topBlockedUrls",
          "topBlockedIps"
        ]
      },
...
    }
  }
}